+++ src/contexts/Farming/Context.ts
@@ -5,10 +5,6 @@ import { ContextValues } from "./types";
 const Context = createContext<ContextValues>({
   farmingStartTime: [1600545500000],
   setConfirmTxModalIsOpen: () => {},
-  onHarvest: () => {},
-  onRedeem: () => {},
-  onStake: () => {},
-  onUnstake: () => {},
 });
 
 export default Context;