+++ src/views/Farm/components/Harvest/Harvest.tsx
@@ -11,14 +11,19 @@ import { AvailableFarms } from "farms/AvailableFarms";
 import LoadingButton from "@mui/lab/LoadingButton";
 import AttachMoneyIcon from "@mui/icons-material/AttachMoney";
 
+import useUbiq from "hooks/useUbiq";
+import { harvest } from "ubiq-sdk/utils";
+
 interface HarvestProps {
   farmKey: number;
 }
 
 const Harvest: React.FC<HarvestProps> = ({ farmKey }) => {
+  const [isHarvesting, setisHarvesting] = useState(false);
   const [earnedBalance, setEarnedBalance] = useState<number>(0);
-  const { status } = useWallet();
-  const { earnedBalances, isHarvesting, onHarvest } = useFarming();
+  const { status, account } = useWallet();
+  const { earnedBalances, setConfirmTxModalIsOpen } = useFarming();
+  const ubiq = useUbiq();
 
   const formattedEarnedBalance = useCallback(async () => {
     if (earnedBalances && bnToDec(earnedBalances[farmKey]) > 0) {
@@ -27,6 +32,27 @@ const Harvest: React.FC<HarvestProps> = ({ farmKey }) => {
       setEarnedBalance(0);
     }
   }, [earnedBalances, farmKey]);
+
+  const handleHarvest = useCallback(async () => {
+    if (!ubiq) return;
+    setConfirmTxModalIsOpen(true);
+    setisHarvesting(true);
+    await harvest(ubiq, account, ubiq.contracts.pools[farmKey], (txHash: string) => {
+      if (txHash === "") {
+        setConfirmTxModalIsOpen(false);
+        setisHarvesting(false);
+      }
+    }).catch((err) => {
+      if (err.code === 4001) {
+        console.log("Wallet: User cancelled");
+      } else {
+        console.log("Error caught:", err);
+      }
+    });
+
+    setConfirmTxModalIsOpen(false);
+    setisHarvesting(false);
+  }, [account, setConfirmTxModalIsOpen, setisHarvesting, ubiq, farmKey]);
 
   useEffect(() => {
     formattedEarnedBalance();
@@ -38,12 +64,12 @@ const Harvest: React.FC<HarvestProps> = ({ farmKey }) => {
     if (status !== "connected") {
       return <p>Connect Wallet</p>;
     }
-    if (isHarvesting !== undefined && isHarvesting[farmKey] === false) {
+    if (isHarvesting === false) {
       return (
         <LoadingButton
           disabled={earnedBalance <= 0}
           onClick={() => {
-            onHarvest(farmKey);
+            handleHarvest();
           }}
           endIcon={<AttachMoneyIcon />}
           loading={false}
@@ -56,10 +82,10 @@ const Harvest: React.FC<HarvestProps> = ({ farmKey }) => {
         </LoadingButton>
       );
     }
-    if (isHarvesting !== undefined && isHarvesting[farmKey] === true) {
+    if (isHarvesting === true) {
       return <p>Harvesting...</p>;
     }
-  }, [status, isHarvesting, earnedBalance, onHarvest, farmKey]);
+  }, [status, isHarvesting, earnedBalance, handleHarvest]);
 
   return (
     <div style={{ padding: "10px 5px 10px 5px" }}>