+++ src/views/Help/Help.tsx
@@ -36,7 +36,7 @@ const Help: React.FC = () => {
         <StyledUl>
           <ListItem legend="Step 1: Setup your wallet">
             Tentacle.Finance supports Web3 wallets such as Metamask and Sparrow.
-            <Button  onClick={openURL("https://ubiqsmart.com/en/wallets")} variant="outlined">
+            <Button onClick={openURL("https://ubiqsmart.com/en/wallets")} variant="outlined">
               Learn how to setup a wallet
             </Button>
           </ListItem>
@@ -47,9 +47,9 @@ const Help: React.FC = () => {
 
           <ListItem legend="Step 3: Get some tokens">
             Obtain UBQ and INK balance in your wallet
-            <ButtonGroup sx={{margin: "10px"}}>
+            <ButtonGroup sx={{ margin: "10px" }}>
               <h3>Obtain UBQ at:</h3>
-              <Button sx={{marginLeft: "10px"}} onClick={openURL("https://bittrex.com/Market/Index?MarketName=USDT-UBQ")} variant="outlined">
+              <Button sx={{ marginLeft: "10px" }} onClick={openURL("https://bittrex.com/Market/Index?MarketName=USDT-UBQ")} variant="outlined">
                 Bittrex
               </Button>
               <Button onClick={openURL("https://dovewallet.com/en/trade/spot/ubq-usdt")} variant="outlined">
@@ -61,7 +61,7 @@ const Help: React.FC = () => {
             </ButtonGroup>
             <ButtonGroup>
               <h3>Obtain INK at:</h3>
-              <Button sx={{marginLeft: "10px"}} onClick={openURL("https://shinobi.ubiq.ninja/#/swap?outputCurrency=" + INK)} variant="outlined">
+              <Button sx={{ marginLeft: "10px" }} onClick={openURL("https://shinobi.ubiq.ninja/#/swap?outputCurrency=" + INK)} variant="outlined">
                 Shinobi Swap
               </Button>
               <Button onClick={addInkToWallet} variant="outlined">
@@ -79,7 +79,7 @@ const Help: React.FC = () => {
 
           <ListItem legend="Step 5: Stake LP tokens">
             Stake your LP tokens on the{" "}
-            <Button sx={{marginLeft: "10px"}} onClick={openURL("/farm")} variant="outlined">
+            <Button sx={{ marginLeft: "10px" }} onClick={openURL("/farm")} variant="outlined">
               Tentacle.Finance Farm
             </Button>
           </ListItem>